Change to use the same colorMap for the same processes.
