+++ straceprof-js/src/ProcessCanvas.tsx
@@ -8,51 +8,11 @@ interface ProcessCanvasProps {
   height: number;
   title: string;
   thresholdToShowProcess: number;
+  colorMap: Record<string, string>;
   onHover?: (
     process: Process | null,
     position: { x: number; y: number } | null
   ) => void;
-}
-
-/**
- * Generate a color map for processes based on program names
- * Similar to the Python implementation's gen_color_map function
- */
-function generateColorMap(processes: Process[]): Record<string, string> {
-  // Count total duration for each program
-  const histogram: Record<string, number> = {};
-
-  for (const process of processes) {
-    const programName = process.program.split('/').pop() || process.program;
-    if (histogram[programName]) {
-      histogram[programName] += process.endTime - process.startTime;
-    } else {
-      histogram[programName] = process.endTime - process.startTime;
-    }
-  }
-
-  // Sort programs by total duration
-  const coloredPrograms = Object.entries(histogram);
-  coloredPrograms.sort((a, b) => b[1] - a[1]);
-
-  // Assign colors to programs
-  const colorList = [
-    '#FF0000', // red
-    '#FFA500', // orange
-    '#FFFF00', // yellow
-    '#FF00FF', // magenta
-    '#800080', // purple
-    '#0000FF', // blue
-    '#00FFFF', // cyan
-    '#008000', // green
-  ];
-
-  const colorMap: Record<string, string> = {};
-  for (let i = 0; i < Math.min(coloredPrograms.length, colorList.length); i++) {
-    colorMap[coloredPrograms[i][0]] = colorList[i];
-  }
-
-  return colorMap;
 }
 
 /**
@@ -104,6 +64,7 @@ const ProcessCanvas: React.FC<ProcessCanvasProps> = ({
   height,
   title,
   thresholdToShowProcess,
+  colorMap,
   onHover,
 }) => {
   const canvasRef = useRef<HTMLCanvasElement>(null);
@@ -204,8 +165,7 @@ const ProcessCanvas: React.FC<ProcessCanvasProps> = ({
       canvas.height = requiredHeight;
     }
 
-    // Generate color map
-    const colorMap = generateColorMap(filteredProcesses);
+    // Use the provided color map from props
 
     // Draw title
     ctx.font = '16px Arial';
@@ -284,6 +244,7 @@ const ProcessCanvas: React.FC<ProcessCanvasProps> = ({
     height,
     thresholdToShowProcess,
     title,
+    colorMap,
   ]);
 
   return (
